# contracts/openapi/horizon-api.yaml
openapi: 3.1.0
info:
  title: Horizon Public API
  version: 1.0.0
  description: |
    Публичный API платформы Horizon — обнаружение зарождающихся научно-технологических
    трендов (слабых сигналов) в открытых источниках.

    Контракт **first-class**: спецификация является источником истины, реализация проверяется
    контрактными тестами (`spring-boot-starter-web` + `swagger-request-validator`, Schemathesis для Python).

    Все ошибки — RFC 9457 `application/problem+json` (ADR-0013).
    Входа нет: организация одна, и API открыт так же, как интерфейс. Служебные вызовы между
    сервисами защищены общим секретом (`X-Internal-Token`) и в этот контракт не входят.
  contact:
    name: Horizon Team
    url: https://horizon.dev
  license:
    name: Proprietary
servers:
  - url: http://localhost:8080
    description: Локальная среда (gateway)
  - url: https://horizon.example.bank
    description: Промышленная среда
...
paths:
...
  /api/v1/saved-domains/refresh:
    post:
      tags: [SavedDomains]
      summary: Поставить на пересчёт все отслеживаемые направления
      description: >-
        Радар полезен ровно настолько, насколько свеж. Обновлять двадцать направлений по одному —
        работа, которую никто не делает, и экран начинает показывать позапрошлый квартал с
        уверенным видом.


        Отвечает 200 со строкой на каждое направление даже если не поставлено ни одно: три исхода
        из четырёх не являются отказом, а HTTP-ошибка заставила бы клиента гадать, что произошло с
        каждым направлением. Пакет подчиняется той же квоте и тому же переиспользованию свежего
        результата, что и одиночный запуск, — иначе кнопка стала бы способом обойти лимит.
      operationId: refreshRadar
      responses:
        '200':
          description: Исход по каждому направлению
          content:
            application/json:
              schema:
                type: array
                items: { $ref: '#/components/schemas/RadarRefresh' }
        '404': { $ref: '#/components/responses/NotFoundProblem' }
...
  /api/v1/saved-domains/{id}/seen:
    post:
      tags: [SavedDomains]
      summary: Отметить направление просмотренным
      description: >-
        Отдельное действие, а не побочный эффект загрузки радара: радар открывают, чтобы решить, куда
        смотреть, и пометить в этот момент всё просмотренным значило бы стереть ответ на вопрос, ради
        которого пришли.


        Отметка личная: визит коллеги не гасит чужую новизну.
      operationId: markDirectionSeen
      parameters:
        - { name: id, in: path, required: true, schema: { type: string, format: uuid } }
      responses:
        '204': { description: Отмечено }
        '404': { $ref: '#/components/responses/NotFoundProblem' }
...
  /api/v1/saved-domains/overlap:
    get:
      tags: [SavedDomains]
      summary: Темы, встречающиеся сразу в нескольких направлениях
      description: >-
        Тема, найденная в одном направлении, — наблюдение. Та же тема, независимо всплывшая в двух
        несмежных направлениях, — конвергенция: признак, что технология выходит за пределы своей
        ниши. Методология этого не видит по построению: индикаторы считаются внутри корпуса одного
        запроса.


        Баллов здесь нет намеренно. Балл нормирован внутри своего корпуса (§12 методологии), у
        баллов разных направлений нет общей шкалы, и число, показанное рядом с другим, будет
        сравнено читателем независимо от подписи. Сопоставляется состав и позиция.


        Пустой список при `directionsCompared` меньше двух означает «сравнивать пока не с чем», а не
        «пересечений нет»; это разные ответы, и клиент обязан их различать.
      operationId: getDirectionOverlap
      responses:
        '200':
          description: Пересечение направлений портфеля
          content:
            application/json:
              schema: { $ref: '#/components/schemas/DirectionOverlap' }
        '404': { $ref: '#/components/responses/NotFoundProblem' }
...
  /api/v1/saved-domains:
    get:
      tags: [SavedDomains]
      summary: Сохранённые направления
      operationId: listSavedDomains
      responses:
        '200':
          description: Список
          content:
            application/json:
              schema:
                type: array
                items: { $ref: '#/components/schemas/SavedDomain' }
    post:
      tags: [SavedDomains]
      summary: Сохранить направление
      operationId: saveDomain
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/SaveDomainRequest' }
      responses:
        '201':
          description: Сохранено
          content:
            application/json:
              schema: { $ref: '#/components/schemas/SavedDomain' }
        '409': { $ref: '#/components/responses/ConflictProblem' }
  /api/v1/saved-domains/{id}:
    delete:
      tags: [SavedDomains]
      summary: Удалить сохранённое направление
      operationId: deleteSavedDomain
      parameters:
        - { name: id, in: path, required: true, schema: { type: string, format: uuid } }
      responses:
        '204': { description: Удалено }
        '404': { $ref: '#/components/responses/NotFoundProblem' }

  # ───────────────────────────── Features ─────────────────────────────
...
components:
...
  responses:
...
    NotFoundProblem:
      description: Не найдено
      content:
        application/problem+json:
          schema: { $ref: '#/components/schemas/Problem' }
    ConflictProblem:
      description: Конфликт состояния
      content:
        application/problem+json:
          schema: { $ref: '#/components/schemas/Problem' }
...
  schemas:
...
    Problem:
      type: object
      description: RFC 9457 Problem Details
      required: [type, title, status]
      properties:
        type: { type: string, format: uri, examples: ['https://horizon.dev/problems/validation-error'] }
        title: { type: string }
        status: { type: integer }
        detail: { type: string }
        instance: { type: string }
        traceId: { type: string }
        timestamp: { type: string, format: date-time }
        retryable: { type: boolean }
        errors:
          type: array
          items:
            type: object
            required: [field, message]
            properties:
              field: { type: string }
              message: { type: string }
              rejectedValue: {}
...
    SourceClass:
      type: string
      enum: [PREPRINT, JOURNAL_ARTICLE, PATENT, CODE_REPOSITORY, NEWS, ANALYST_REPORT, STANDARD]
...
    AnalysisParameters:
      type: object
      properties:
        topN: { type: integer, minimum: 5, maximum: 50, default: 15 }
        yearsWindow: { type: integer, minimum: 3, maximum: 15, default: 7 }
        sourceClasses:
          type: array
          items: { $ref: '#/components/schemas/SourceClass' }
          description: 'Пустой список — все доступные классы'
        minConfidence: { type: number, minimum: 0, maximum: 1, default: 0 }
        includeMature: { type: boolean, default: false }
        mode:
          type: string
          enum: [fast, quality]
          default: fast
          description: >-
            Режим анализа. `fast` — ответ за 20 минут (SLA ТЗ): глубокое исследование получает
            5,5 минуты. `quality` — до 30–40 минут: глубокое исследование читает втрое больше
            страниц за 20 минут, дедлайн запроса — 40 минут.


            Часть идентичности вопроса: быстрый и качественный ответы на одно направление — два
            разных ответа, и готовый отчёт одного режима не переиспользуется как ответ другого.
            Незнакомое имя — 400 с перечнем известных.
...
    SaveDomainRequest:
      type: object
      required: [query]
      properties:
        query: { type: string, minLength: 3, maxLength: 200 }
        parameters: { $ref: '#/components/schemas/AnalysisParameters' }
    SavedDomain:
      type: object
      required: [id, query, createdAt]
      properties:
        id: { type: string, format: uuid }
        query: { type: string }
        normalizedQuery: { type: string }
        parameters: { $ref: '#/components/schemas/AnalysisParameters' }
        lastReportId: { type: [string, 'null'], format: uuid }
        createdAt: { type: string, format: date-time }
    RadarRefresh:
      type: object
      required: [savedDomainId, query, outcome]
      properties:
        savedDomainId: { type: string, format: uuid }
        query: { type: string }
        outcome:
          type: string
          description: >-
            `accepted` — поставлено в очередь. `already-running` — анализ по этому направлению уже
            идёт, второй не запущен и квота не списана. `quota-exceeded` — часовой лимит исчерпан.
            `failed` — отказ по собственной причине направления.


            Исхода «переиспользован готовый результат» здесь нет: пакетный пересчёт принудителен —
            кнопка заводилась ради обновления, и отвечать «уже свежее» на все направления сразу
            значит не делать ничего. Каждое направление расходует слот квоты.


            Строка, а не закрытый enum: исход показывается пользователю, и появление нового не
            должно ломать уже собранных клиентов.
        requestId:
          type: string
          format: uuid
          description: >-
            Есть для `accepted` и `already-running` — по нему клиент следит за анализом.
        reason:
          type: string
          description: Заполнено для `quota-exceeded` и `failed`.
...
    DirectionOverlap:
      type: object
      required: [topics, directionsCompared, directionsSkipped]
      properties:
        topics:
          type: array
          items: { $ref: '#/components/schemas/SharedTopic' }
        directionsCompared:
          type: integer
          description: Сколько направлений участвовало — то есть имеет отчёт.
        directionsSkipped:
          type: integer
          description: >-
            Сколько отслеживаемых направлений не участвовало, потому что анализ по ним не
            запускался. Названо числом, а не молчанием: иначе «пересечений нет» читалось бы как
            утверждение о предметной области.
    SharedTopic:
      type: object
      required: [trendKey, title, appearances]
      properties:
        trendKey:
          type: string
          description: >-
            Идентичность темы по методологии — тот же ключ, что у обратной связи, дельты и
            трассировки. Сопоставление идёт по нему, а не по заголовку.
        title:
          type: string
          description: Заголовок из того направления, где тема стоит выше всего.
        appearances:
          type: array
          minItems: 2
          items: { $ref: '#/components/schemas/TopicAppearance' }
    TopicAppearance:
      type: object
      required: [savedDomainId, query, reportId, rank]
      properties:
        savedDomainId: { type: string, format: uuid }
        query: { type: string }
        reportId: { type: string, format: uuid }
        rank:
          type: integer
          description: >-
            Место темы в этом направлении. Вторая позиция в одном и пятнадцатая в другом — другое
            утверждение, чем «есть в обоих».
...